requirements.txt
conan>=0.19.0
pytest==3.0.3

